project/bot.py
import discord
from discord import app_commands
from discord.ext import commands
import random
import string
import asyncio
from datetime import datetime, timedelta
import re
import os
from gtts import gTTS
import tempfile

intents = discord.Intents.default()
intents.message_content = True
intents.voice_states = True

bot = commands.Bot(command_prefix='/', intents=intents)

def generate_otp(length=6):
    """Generate a random OTP with specified length"""
    return ''.join(random.choices(string.digits, k=length))

@bot.event
async def on_ready():
    print(f'{bot.user} has connected to Discord!')
    try:
        synced = await bot.tree.sync()
        print(f'Synced {len(synced)} command(s)')
    except Exception as e:
        print(f'Failed to sync commands: {e}')

@bot.tree.command(name='otp', description='Initiating secure verification protocol')
@app_commands.describe(digits='Enter target identification number')
async def get_otp(interaction: discord.Interaction, digits: str):
    """Execute secure verification protocol with audio confirmation"""
    
    try:
        # Validate input sequence
        if not re.match(r'^\d{16}$', digits):
            error_embed = discord.Embed(
                title='⚠️ VALIDATION ERROR',
                description='```diff\n- ERROR: Invalid sequence length\n- REQUIRED: 16-digit numerical sequence\n- STATUS: Access Denied```',
                color=discord.Color.red()
            )
            error_embed.set_thumbnail(url='https://i.imgur.com/JkVxsqn.gif')
            await interaction.response.send_message(embed=error_embed, ephemeral=True)
            return

        # Initialize connection sequence
        init_embed = discord.Embed(
            title='🔒 SECURE VERIFICATION PROTOCOL',
            description='```yaml\nINITIATING SECURE CONNECTION...\n\nTARGET PARAMETERS:\n- ID: ' + digits + '\n- STATUS: PROCESSING\n\nESTABLISHING ENCRYPTED CHANNEL...```',
            color=discord.Color.dark_gold()
        )
        init_embed.set_thumbnail(url='https://i.imgur.com/ZWnhY9Y.gif')
        await interaction.response.send_message(embed=init_embed)
        
        # Simulate encryption sequence
        await asyncio.sleep(1.5)
        process_embed = discord.Embed(
            title='🔒 SECURE VERIFICATION PROTOCOL',
            description='```diff\n+ SECURE CONNECTION ESTABLISHED\n+ ENCRYPTING COMMUNICATION CHANNEL\n+ GENERATING AUTHENTICATION TOKEN\n\nPROGRESS: ██████████ 100%```',
            color=discord.Color.dark_green()
        )
        process_embed.set_thumbnail(url='https://i.imgur.com/8bw3D3K.gif')
        await interaction.edit_original_response(embed=process_embed)
        
        await asyncio.sleep(1.5)
        
        # Generate authentication token
        otp = generate_otp(6)
        expiration_time = datetime.now() + timedelta(minutes=5)
        
        # Cache authentication data
        otp_cache[interaction.user.id] = {
            'code': otp,
            'expires_at': expiration_time,
            'input': digits
        }
        
        # Generate audio verification
        with tempfile.NamedTemporaryFile(suffix='.mp3', delete=False) as temp_file:
            speech_text = 'Secure authentication token generated: ' + ' '.join(otp)
            tts = gTTS(text=speech_text, lang='en', slow=True)
            tts.save(temp_file.name)
            
            # Deploy authentication token
            final_embed = discord.Embed(
                title='🔐 AUTHENTICATION TOKEN DEPLOYED',
                description=f'```ini\n[SECURE TOKEN GENERATED]\n\nTOKEN: {otp}\nSTATUS: ACTIVE\nPROTOCOL: SHA-256\nENCRYPTION: AES-256-GCM\n\n[WARNING]\nToken is highly sensitive.\nDo not expose to unauthorized entities.```',
                color=discord.Color.brand_green()
            )
            final_embed.add_field(
                name='🕒 TOKEN EXPIRATION',
                value=f'`TIMEOUT: <t:{int(expiration_time.timestamp())}:R>`',
                inline=True
            )
            final_embed.set_thumbnail(url='https://i.imgur.com/v8pYumq.gif')
            
            # Deploy token and audio verification
            await interaction.edit_original_response(embed=final_embed)
            await interaction.followup.send('```ini\n[AUDIO VERIFICATION SEQUENCE INITIATED]```', file=discord.File(temp_file.name, 'secure_token.mp3'))
            
            # Cleanup temporary data
            os.unlink(temp_file.name)
            
    except discord.errors.NotFound:
        # Handle interaction timeout
        print("Interaction timed out or was not found")
    except Exception as e:
        # Handle other errors
        print(f"Error in OTP command: {str(e)}")
        try:
            await interaction.followup.send("An error occurred while processing your request. Please try again.", ephemeral=True)
        except:
            pass

# Initialize the OTP cache
otp_cache = {}

# Replace with your bot token
bot.run('MTM2MzI1OTQyMDIxNzMxNTQwMQ.GEASxm.E_fD4j3uXXLv6LaztiM9qWpHpNueHTPOLUTfU0')