project/requirements.txt
discord.py>=2.0.0
PyNaCl>=1.4.0